requirements.txt
crewai>=0.1.0
weaviate-client>=4.6.0
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.28.0
pytest>=7.0.0
python-dateutil>=2.8.2
paho-mqtt>=1.6.1
email-validator>=1.3.0
